extract points history to its own component
