+++ src/components/Global/LevelsCard/LevelsCard.module.css
@@ -45,37 +45,7 @@ max-width: 100%;
     height: 1px;
     background: var(--dark3)
 }
-.point_history_container{
-    
-    padding: 8px;
-    display: flex;
-    flex-direction: column;
-    gap:8px;
 
-width: 100%;
-  
-
-}
-.points_history_content{
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    border: 1px solid var(--dark3);
-    border-radius: 4px;
-    height: 101px;
-    overflow-y: scroll;
-    padding: 0 8px;
-}
-
-.points_history_header{
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    padding: 0 8px;
-}
-.hide_scroll{
-    overflow-y: hidden;
-}
-
- 
 
 .level_only_container{
 display: flex;
@@ -125,7 +95,7 @@ align-items: center;
 @media (min-width: 768px) {
     
     .main_container{
-width: 405px;
+width: 450px;
 height: 546px;
 margin: auto;
