+++ src/components/Global/LevelsCard/PointsHistoryDisplay/PointsHistoryDisplay.module.css
@@ -1,0 +1,29 @@
+.point_history_container{
+    
+    padding: 8px;
+    display: flex;
+    flex-direction: column;
+    gap:8px;
+
+width: 100%;
+  
+
+}
+.points_history_content{
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    border: 1px solid var(--dark3);
+    border-radius: 4px;
+    height: 101px;
+    overflow-y: scroll;
+    padding: 0 8px;
+}
+
+.points_history_header{
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    padding: 0 8px;
+}
+.hide_scroll{
+    overflow-y: hidden;
+}
